Error handlling refined,  and test files refined
