+++ art_work/file_contents.go
@@ -12,7 +12,7 @@ func FileContents() []string {
 	fileName := utils.GetFile(os.Args)
 	contents, err := os.ReadFile(fileName)
 	if err != nil {
-		utils.ErrorHandler(err)
+		utils.ErrorHandler("seems '" + fileName + "' is missing")
 	}
 
 	var fileContents []string
