+++ ascii/test_art.go
@@ -1,6 +1,6 @@
 package main
 
-const asciiArt = `  _    _          _   _  ` + "\n" +
+const asciiArt = `      _    _          _   _  ` + "\n" +
 	`| |  | |        | | | |         ` + "\n" +
 	`| |__| |   ___  | | | |   ___   ` + "\n" +
 	`|  __  |  / _ \ | | | |  / _ \  ` + "\n" +
